学生提交请假条接口

+++ src/main/java/com/soft1851/springboot/smart/attendance/model/entity/StudentBuilding.java
@@ -22,7 +22,7 @@ public class StudentBuilding {
      */
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer pkId;
+    private Long pkId;
     /**
      * 阿姨id
      */
