+++ src/main/java/com/soft1851/springboot/smart/attendance/model/entity/UserBuilding.java
@@ -22,7 +22,7 @@ public class UserBuilding {
      */
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer pkId;
+    private Long pkId;
     /**
      * 学生id
      */
